isFinished function showed on the header

+++ src/components/Task.tsx
@@ -4,13 +4,14 @@ import style from './Task.module.css'
 import trash from '../assets/trash.png'
 
 
-interface Task {
+export interface TaskProps {
     content: string,
     isFinished?: boolean
     onDeleteTask: (task: string) => void
+    toggleTaskCompletion: (task: string) => void
 }
 
-export function Task({content, isFinished = false, onDeleteTask}: Task){
+export function Task({content, isFinished = false, onDeleteTask, toggleTaskCompletion}: TaskProps){
 
     const [isChecked, setIsChecked] = useState(isFinished)
 
@@ -18,30 +19,39 @@ export function Task({content, isFinished = false, onDeleteTask}: Task){
 
     function handleRadioChange () {
         setIsChecked(!isChecked)
+        toggleTaskCompletion(content);
     }
     function handleDeleteTask () {
-        onDeleteTask(content)
+        // verify box is checked
+        const verifyDelete = confirm('Are you sure you want to delete?')
+
+        if(verifyDelete) {
+            onDeleteTask(content)
+        }
+
+        
     }
 
     return (
-        <div className={style.taskBox}>
-            <div className={style.leftContent}>
-                <input 
-                    type="radio"
-                    className={style.inputRadio}
-                    checked={isChecked}
-                    onChange={handleRadioChange}
-                    />
-                <p>{content}</p>
-            </div>
-            
-            <button 
-                className={style.deleteButton}
-                onClick={handleDeleteTask}
-                >
-                <img src={trash} alt="" />
+        <div className={ isChecked ? style.checked :  style.taskBox} >
+                <div className={style.leftContent}>
+                    <input 
+                        type="radio"
+                        className={style.inputRadio}
+                        checked={isChecked}
+                        onChange={handleRadioChange}
+                        />
+                    <p>{content}</p>
+                </div>
                 
-            </button>
+                <button 
+                    className={style.deleteButton}
+                    onClick={handleDeleteTask}
+                    >
+                    <img src={trash} alt="" />
+                    
+                </button>
+                
         </div>
     )
 }
